app/requirements.txt
asyncpg==0.28.0
aiogram==3.1.1
aiogram-dialog==2.1.0
python-dotenv==1.0.0